Rework top-level sign destruction.
Counterpart to the sign creation rework, this commit uses the new sign
store and sign writer. Sign destruction works "in reverse" compared to
sign creation in the sense that we first remove from the store and then
erase from disk. I suppose we really should erase from disk first since
it's the "golden record", but it would require either re-adding the sign
on errors with erase or doing a lookup and then a removal, which feels a
bit icky.

The sign writer probably needs to be idempotent and print a warning if
it fails to erase due to invalid data or missing entries, just to add a
bit of resilience in the event handler.