Flesh out sign writer.
This commit introduces two new stubs for file system operations and
serialization. At first glance, both of these stubs might look like
excessive abstraction, but it turns out that both responsibilities are a
little more involved when you dig down. File system stuff is always
annoying to test, and serialization and deserialization are a brittle
duality that is probably best tested in isolation.