Flesh out sign serializer.
Implements the equal and serialize methods. There will be a deserialize
method once the world load event handler is implemented.

The equal method specifically only checks world ID and coords (x,y,z),
because that _should_ be a unique key for the signs as two different
signs cannot physically exist in the same location.