Make sign writer erase conflicts on write.
We may run into issues with "duplicate" entries if a sign somehow breaks
outside of the BlockBreakEvent and is re-placed. This means we need some
extra robustness in the writer's invariants for the write procedure. We
want it to be true that every time the procedure has run, the location
in question is unique in the data file.